requirements.txt
pytest==7.4.2
pytest-timeout==2.1.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
mypy-extensions==1.0.0
packaging==23.1
colorama==0.4.6
idna==3.4
iniconfig==2.0.0
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.3.0
requests==2.30.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
playwright~=1.37.0
pydantic==2.3.0